requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
garth>=0.5.16
garminconnect>=0.2.28
fit-tool>=0.9.13
